requirements.txt
selenium==selenium 3.141.0
chromedriver-binary-auto==0.1